# internal/api/v1/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Tailnet API",
    "title": "Tailnet API",
    "termsOfService": "https://github.com/sudosu404/providers/blob/main/LICENSE",
    "contact": {
      "name": "Yusing",
      "url": "https://github.com/sudosu404/providers/issues"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/sudosu404/providers/blob/main/LICENSE"
    },
    "version": "1.0"
  },
  "basePath": "/api/v1",
  "paths": {
    "/agent/create": {
      "post": {
        "description": "Create a new agent and return the docker compose file, encrypted CA and client PEMs\nThe returned PEMs are encrypted with a random key and will be used for verification when adding a new agent",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "agent"
        ],
        "summary": "Create a new agent",
        "parameters": [
          {
            "description": "Request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewAgentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/NewAgentResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "create",
        "operationId": "create"
      }
    },
    "/agent/list": {
      "get": {
        "description": "List agents",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "agent",
          "websocket"
        ],
        "summary": "List agents",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Agent"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "list",
        "operationId": "list"
      }
    },
    "/agent/verify": {
      "post": {
        "description": "Verify a new agent and return the number of routes added",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "agent"
        ],
        "summary": "Verify a new agent",
        "parameters": [
          {
            "description": "Request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VerifyNewAgentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "verify",
        "operationId": "verify"
      }
    },
    "/auth/callback": {
      "post": {
        "description": "Handles the callback from the provider after successful authentication",
        "produces": [
          "text/plain"
        ],
        "tags": [
          "auth"
        ],
        "summary": "Auth Callback",
        "parameters": [
          {
            "description": "Userpass only",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/auth.UserPassAuthCallbackRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Userpass: OK",
            "schema": {
              "type": "string"
            }
          },
          "302": {
            "description": "OIDC: Redirects to home page",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Userpass: invalid request / credentials",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "string"
            }
          }
        },
        "x-id": "callback",
        "operationId": "callback"
      }
    },
...
    "/auth/login": {
      "post": {
        "description": "Initiates the login process by redirecting the user to the provider's login page",
        "produces": [
          "text/plain"
        ],
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "responses": {
          "302": {
            "description": "Redirects to login page or IdP",
            "schema": {
              "type": "string"
            }
          },
          "429": {
            "description": "Too Many Requests",
            "schema": {
              "type": "string"
            }
          }
        },
        "x-id": "login",
        "operationId": "login"
      }
    },
...
    "/cert/info": {
      "get": {
        "description": "Get cert info",
        "produces": [
          "application/json"
        ],
        "tags": [
          "cert"
        ],
        "summary": "Get cert info",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CertInfo"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "info",
        "operationId": "info"
      }
    },
    "/cert/renew": {
      "get": {
        "description": "Renew cert",
        "produces": [
          "text/plain"
        ],
        "tags": [
          "cert",
          "websocket"
        ],
        "summary": "Renew cert",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "renew",
        "operationId": "renew"
      }
    },
    "/docker/container/{id}": {
      "get": {
        "description": "Get container by container id",
        "produces": [
          "application/json"
        ],
        "tags": [
          "docker"
        ],
        "summary": "Get container",
        "parameters": [
          {
            "type": "string",
            "description": "Container ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ContainerResponse"
            }
          },
          "400": {
            "description": "ID is required",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Container not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "container",
        "operationId": "container"
      }
    },
    "/docker/containers": {
      "get": {
        "description": "Get containers",
        "produces": [
          "application/json"
        ],
        "tags": [
          "docker"
        ],
        "summary": "Get containers",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ContainerResponse"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "containers",
        "operationId": "containers"
      }
    },
    "/docker/info": {
      "get": {
        "description": "Get docker info",
        "produces": [
          "application/json"
        ],
        "tags": [
          "docker"
        ],
        "summary": "Get docker info",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ServerInfo"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "info",
        "operationId": "info"
      }
    },
    "/docker/logs/{id}": {
      "get": {
        "description": "Get docker container logs by container id",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "docker",
          "websocket"
        ],
        "summary": "Get docker container logs",
        "parameters": [
          {
            "type": "string",
            "description": "container id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "boolean",
            "description": "show stdout",
            "name": "stdout",
            "in": "query"
          },
          {
            "type": "boolean",
            "description": "show stderr",
            "name": "stderr",
            "in": "query"
          },
          {
            "type": "string",
            "description": "from timestamp",
            "name": "from",
            "in": "query"
          },
          {
            "type": "string",
            "description": "to timestamp",
            "name": "to",
            "in": "query"
          },
          {
            "type": "string",
            "description": "levels",
            "name": "levels",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "server not found or container not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "logs",
        "operationId": "logs"
      }
    },
    "/docker/restart": {
      "post": {
        "description": "Restart container by container id",
        "produces": [
          "application/json"
        ],
        "tags": [
          "docker"
        ],
        "summary": "Restart container",
        "parameters": [
          {
            "description": "Request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dockerapi.StopRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Container not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "restart",
        "operationId": "restart"
      }
    },
    "/docker/start": {
      "post": {
        "description": "Start container by container id",
        "produces": [
          "application/json"
        ],
        "tags": [
          "docker"
        ],
        "summary": "Start container",
        "parameters": [
          {
            "description": "Request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dockerapi.StartRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Container not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "start",
        "operationId": "start"
      }
    },
    "/docker/stop": {
      "post": {
        "description": "Stop container by container id",
        "produces": [
          "application/json"
        ],
        "tags": [
          "docker"
        ],
        "summary": "Stop container",
        "parameters": [
          {
            "description": "Request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dockerapi.StopRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Container not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "stop",
        "operationId": "stop"
      }
    },
    "/favicon": {
      "get": {
        "description": "Get favicon",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "image/svg+xml",
          "image/x-icon",
          "image/png",
          "image/webp"
        ],
        "tags": [
          "v1"
        ],
        "summary": "Get favicon",
        "parameters": [
          {
            "type": "string",
            "description": "URL of the route",
            "name": "url",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Alias of the route",
            "name": "alias",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/homepage.FetchResult"
              }
            }
          },
          "400": {
            "description": "Bad Request: alias is empty or route is not HTTPRoute",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden: unauthorized",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found: route or icon not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error: internal error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "favicon",
        "operationId": "favicon"
      }
    },
    "/file/content": {
      "get": {
        "description": "Get file content",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json",
          "application/godoxy+yaml"
        ],
        "tags": [
          "file"
        ],
        "summary": "Get file content",
        "parameters": [
          {
            "type": "string",
            "format": "filename",
            "name": "filename",
            "in": "query",
            "required": true
          },
          {
            "enum": [
              "config",
              "provider",
              "middleware"
            ],
            "type": "string",
            "x-enum-varnames": [
              "FileTypeConfig",
              "FileTypeProvider",
              "FileTypeMiddleware"
            ],
            "name": "type",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "get",
        "operationId": "get"
      },
      "put": {
        "description": "Set file content",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "file"
        ],
        "summary": "Set file content",
        "parameters": [
          {
            "enum": [
              "config",
              "provider",
              "middleware"
            ],
            "type": "string",
            "description": "Type",
            "name": "type",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "Filename",
            "name": "filename",
            "in": "query",
            "required": true
          },
          {
            "description": "File",
            "name": "file",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "set",
        "operationId": "set"
      }
    },
    "/file/list": {
      "get": {
        "description": "List files",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "file"
        ],
        "summary": "List files",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ListFilesResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "list",
        "operationId": "list"
      }
    },
    "/file/validate": {
      "post": {
        "description": "Validate file",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "file"
        ],
        "summary": "Validate file",
        "parameters": [
          {
            "enum": [
              "config",
              "provider",
              "middleware"
            ],
            "type": "string",
            "description": "Type",
            "name": "type",
            "in": "query",
            "required": true
          },
          {
            "description": "File content",
            "name": "file",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File validated",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "417": {
            "description": "Validation failed",
            "schema": {}
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "validate",
        "operationId": "validate"
      }
    },
...
    "/homepage/categories": {
      "get": {
        "description": "List homepage categories",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "homepage"
        ],
        "summary": "List homepage categories",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "categories",
        "operationId": "categories"
      }
    },
    "/homepage/item_click": {
      "post": {
        "description": "Increment item click.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "homepage"
        ],
        "summary": "Increment item click",
        "parameters": [
          {
            "type": "string",
            "name": "which",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "item-click",
        "operationId": "item-click"
      }
    },
...
    "/reload": {
      "post": {
        "description": "Reload config",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "v1"
        ],
        "summary": "Reload config",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "x-id": "reload",
        "operationId": "reload"
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-nullable": true
        },
        "message": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
    "NewAgentRequest": {
      "type": "object",
      "required": [
        "host",
        "name",
        "port",
        "type"
      ],
      "properties": {
        "container_runtime": {
          "default": "docker",
          "enum": [
            "docker",
            "podman"
          ],
          "allOf": [
            {
              "$ref": "#/definitions/agent.ContainerRuntime"
            }
          ]
        },
        "host": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "name": {
          "type": "string",
          "x-nullable": false,
          "x-omitempty": false
        },
        "nightly": {
          "type": "boolean"
        },
        "port": {
          "type": "integer",
          "maximum": 65535,
          "minimum": 1,
          "x-nullable": false,
          "x-omitempty": false
        },
        "type": {
          "type": "string",
          "enum": [
            "docker",
            "system"
          ],
          "x-nullable": false,
          "x-omitempty": false
        }
      },
      "x-nullable": false,
      "x-omitempty": false
    },
...
  },
...
}